pyproject.toml
[tool.poetry]
name = "teste"
version = "0.1.0"
description = ""
authors = ["Maria Eduarda Neves Correia <mariaae.correia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
djangorestframework-simplejwt = "^4.6.0"
drf-spectacular = "^0.14.0"
django-cors-headers = "^3.7.0"
django-environ = "^0.4.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
